+++ inc/utils/Utils.hpp
@@ -6,7 +6,7 @@
 /*   By: fdi-cecc <fdi-cecc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/30 15:21:29 by fdi-cecc          #+#    #+#             */
-/*   Updated: 2025/06/30 15:34:17 by fdi-cecc         ###   ########.fr       */
+/*   Updated: 2025/07/01 12:47:20 by fdi-cecc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,6 @@
 #include <string>
 #include <iostream>
 
-std::string utilsTimestamp();
+const std::string utilsTimestamp();
 
 #endif
